requirements.txt
functions-framework==3.*
langchain
pinecone
langchain_openai
google-cloud-bigquery
google-cloud-storage
google-cloud
langchain_pinecone